Move the interface for artifact storage code to interface/artifacts/storage.py
