+++ wandb/sdk/interface/artifacts/__init__.py
@@ -1,9 +1,3 @@
-from .artifacts import (
-    Artifact,
-    ArtifactEntry,
-    ArtifactManifest,
-    StorageHandler,
-    StorageLayout,
-    StoragePolicy,
-)
+from .artifacts import Artifact, ArtifactEntry, ArtifactManifest
 from .cache import ArtifactsCache, get_artifacts_cache
+from .storage import StorageHandler, StorageLayout, StoragePolicy
